ohca_frontend/requirements.txt
streamlit
folium
streamlit-folium
requests==2.32.3
pandas==2.2.3
python-dotenv==1.0.1
shapely